# Fav8_Backend/openapi/stations.yml
# Info
openapi: 3.0.2
info:
  title: Radio Stations SDK
  description: Experimental description of the Radio Stations SDK.
  version: 1.0.0

# Servers
servers:
  - url: http://localhost:8080/api
    description: local

tags:
  - name: Radio Stations
    description: Radio Stations APIs

# Paths
paths:
  # Server status
  /stations:
    get:
      operationId: getAllRadioStations
      tags:
        - RadioStations
      summary: This is a summary
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          description: The id to consume.
      description: This is a description
      responses:
        "200":
          description: This means ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Station"
  #   post:
  #     parameters:
  #       - in: query
  #         name: station
  #         description: New station
  #         schema:
  #           type: object
  #           properties:
  #             stationName:
  #               type: string
  #             stationURL:
  #               type: string
  #     responses:
  #       201:
  #         description: Created
  # /stations/{id}:
  #   get:
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         type: string
  #         description: The station ID.
  #   description: Get a station by id
  #   responses:
  #     200:
  #       description: Returns the requested stations
  #   patch:
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         type: string
  #         description: The station ID.
  #       - in: body
  #         name: station
  #         description: Update station
  #         schema:
  #           type: object
  #           properties:
  #             stationName:
  #               type: string
  #             stationURL:
  #               type: string
  #     responses:
  #       201:
  #         description: Created
  #   delete:
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         type: string
  #         description: The station ID.
  #     description: Delete a station by id
  #     responses:
  #       200:
  #         description: Returns the requested stations

# DTOs
components:
  schemas:
    # Radio Stations List DTO
    Station:
      description: This is the radio station model
      required:
        - id
        - stationName
        - stationUrl
      type: object
      properties:
        id:
          description: The id of the radio station
          type: string
        stationName:
          description: The name of the radio station
          type: string
        stationUrl:
          description: The streaming url of the radio station
          type: string