# Fav8_Backend/src/main/java/com/woqomoqo/fav8/sqldata/StationController.java
 package com.woqomoqo.fav8.sqldata;

 import com.woqomoqo.radiostations.api.RadioStationsApi;
// import com.woqomoqo.radiostations.api.model.Station;
 import com.woqomoqo.fav8.sqldata.model.Station;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;

 import org.springframework.web.bind.annotation.*;


 // @Service
 @Controller // This means that this class is a Controller
 @RequestMapping(path = "/api") // This means URL's start with /demo (after Application path)
 public class StationController implements RadioStationsApi {


  @Autowired // This means to get the bean called stationRepository which is auto-generated by Spring, we will use it to handle the data

  private StationRepository stationRepository;


//   private final StationRepository stationRepository;
//
//  public StationController(StationRepository stationRepository) {
//   this.stationRepository = stationRepository;
//  }

 @PostMapping(path = "/add") // Map ONLY POST Requests
 public @ResponseBody String addNewStation(@RequestParam String stationName,
 @RequestParam String stationUrl) {
 // @ResponseBody means the returned String is the response, not a view name
 // @RequestParam means it is a parameter from the GET or POST request

 Station n = new Station();
 n.setStationName(stationName);
 n.setStationUrl(stationUrl);
 stationRepository.save(n);
 return "Saved";
 }

 @GetMapping(path = "/stations")
// @CrossOrigin(origins = "http://localhost:4200")
// @ApiOperation("Returns list of radio stations")
 public @ResponseBody Iterable<Station> getAllRadioStations() {
 // This returns a JSON or XML with the stations
 return stationRepository.findAll();
 }
 }